+++ tests/form-level.spec.ts
@@ -525,7 +525,8 @@ test.describe('Home page', () => {
           responseStub.id = '1111';
           responseStub.meta = {
             versionId: '1',
-            lastUpdated: '2020-02-22T22:22:22.222-00:00'
+            lastUpdated: '2020-02-22T22:22:22.222-00:00',
+            source: '#3JO34SP5oR7bcwTb'
           };
 
           await PWUtils.uploadFile(page, `${testConfig.fixtureFile}`);
@@ -546,8 +547,20 @@ test.describe('Home page', () => {
 
           const created = await PWUtils.getFHIRServerResponse(page, 'Create a new questionnaire on a FHIR server...', testConfig.serverBaseUrl);
           expect(created).toEqual(responseStub);
+          const createdQuestionnaire = await PWUtils.getQuestionnaireJSONWithoutUI(page, testConfig.version);
+          expect(createdQuestionnaire.id).toBe(responseStub.id);
+          expect(createdQuestionnaire.meta.versionId).toBe(responseStub.meta.versionId);
+          expect(createdQuestionnaire.meta.lastUpdated).toBe(responseStub.meta.lastUpdated);
+          if(testConfig.version !== 'STU3') {
+            expect(createdQuestionnaire.meta.source).toBe(responseStub.meta.source);
+          }
 
           responseStub.title = 'Modified title';
+          responseStub.meta = {
+            versionId: '2',
+            lastUpdated: '2020-02-23T22:22:22.222-00:00',
+            source: '#updatedSource'
+          };
           const titleField = await PWUtils.getByLabel(page, 'lfb-form-fields', 'Title');
           await titleField.clear();
           await titleField.fill(responseStub.title);
@@ -567,6 +580,12 @@ test.describe('Home page', () => {
 
           const updated = await PWUtils.getFHIRServerResponse(page, 'Update the questionnaire on the server');
           expect(updated).toEqual(responseStub);
+          const updatedQuestionnaire = await PWUtils.getQuestionnaireJSONWithoutUI(page, testConfig.version);
+          expect(updatedQuestionnaire.meta.versionId).toBe(responseStub.meta.versionId);
+          expect(updatedQuestionnaire.meta.lastUpdated).toBe(responseStub.meta.lastUpdated);
+          if(testConfig.version !== 'STU3') {
+            expect(updatedQuestionnaire.meta.source).toBe(responseStub.meta.source);
+          }
         });
       }
     });